航标站维护保养计划，航标巡检计划

+++ nav-data-sync/src/main/java/top/hting/entity/oracle/paln/TblPlanMark.java
@@ -29,10 +29,10 @@ public class TblPlanMark extends OBase implements Serializable {
     private String planNumber;
 
     @Excel(name = "annual")
-    private Short annual;
+    private Integer annual;
 
     @Excel(name = "monthly")
-    private Short monthly;
+    private Integer monthly;
 
     @Excel(name = "userId")
     private String userId;
